finished workshop page layout, added create workshop button, get-user function
