+++ src/api/post/post-login.js
@@ -1,6 +1,6 @@
 async function postLogin(username, password) {
     const url = `${import.meta.env.VITE_API_URL}/api-token-auth/`;
-    console.log(username, password)
+
     const response = await fetch(url, {
         method: 'POST',
         headers: {
