+++ src/components/AuthProvider.jsx
@@ -7,7 +7,9 @@ export const AuthProvider = (props) => {
         token: window.localStorage.getItem('token'),
         firstName: window.localStorage.getItem('first_name'),
         userId: window.localStorage.getItem('user_id'),
-    })
+        isSuper: window.localStorage.getItem('is_superuser'),
+        isAdmin: window.localStorage.getItem('is_staff'),
+    });
 
     return (
         <AuthContext.Provider value={{auth, setAuth}}>
